Move `GetSwapAngle` to `RawTouchAim`
